set up test structure for data saving
