+++ minialoha/utils/data_utils.py
@@ -1,0 +1,30 @@
+import time
+
+import h5py
+
+
+def save_to_hdf5(data_dict, dataset_path, camera_names, max_timesteps):
+    # HDF5
+    t0 = time.time()
+    with h5py.File(dataset_path + ".hdf5", "w", rdcc_nbytes=1024**2 * 2) as root:
+        root.attrs["sim"] = False
+        obs = root.create_group("observations")
+        image = obs.create_group("images")
+        for cam_name in camera_names:
+            _ = image.create_dataset(
+                cam_name,
+                (max_timesteps, 480, 640, 3),
+                dtype="uint8",
+                chunks=(1, 480, 640, 3),
+            )
+            # compression='gzip',compression_opts=2,)
+            # compression=32001, compression_opts=(0, 0, 0, 0, 9, 1, 1), shuffle=False)
+        _ = obs.create_dataset("qpos", (max_timesteps, 14))
+        _ = obs.create_dataset("qvel", (max_timesteps, 14))
+        _ = obs.create_dataset("effort", (max_timesteps, 14))
+        _ = root.create_dataset("action", (max_timesteps, 14))
+
+        for name, array in data_dict.items():
+            dataset = root[name]
+            dataset[...] = array
+    print(f"Saving: {time.time() - t0:.1f} secs")
