+++ tests/MinialohaExampleDataDict.txt
@@ -1,0 +1,9 @@
+data_dict
+{       '/observations/qpos': [array([ 3497.        ,  1893.        ,  1121.        ,   903.        ,
+                2058.        , 51969.23232323]), array([ 3485.,  1886.,  1121.,   906.,  2065., 52146.]), array([ 3412.        ,  1863.        ,  1123.        ,   942.        ,
+                2085.        , 52651.05050505])],
+        '/observations/qvel': [],
+        '/observations/effort': [],
+        '/action': [array([2762, 1818, 1127, 1073, 2120], dtype=int32), array([2762, 1818, 1127, 1073, 2120], dtype=int32), array([2762, 1818, 1127, 1073, 2120], 
+                dtype=int32)]
+}
